requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
bcrypt==4.3.0
pyotp==2.9.0
qrcode[pil]==8.0
httpx==0.27.2
python-multipart==0.0.12
anthropic==0.40.0
pypdf==5.1.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
pydantic[email]
python-frontmatter==1.1.0
Pillow==12.1.1
